Improved logging and api responses on exceptions

+++ cheeseria/cheeseria.Server/Services/CheeseService.cs
@@ -4,15 +4,22 @@ namespace CheeseApp.Server.Services
 {
     public class CheeseService : ICheeseService
     {
+        ILogger _logger;
+
         ICheeseRepository _cheeseRepository;
         public CheeseService(ILogger<CheeseService> logger, ICheeseRepository cheeseRepository )
         {
+            _logger = logger;
             _cheeseRepository = cheeseRepository;
         }
 
         public async Task<CheeseDTO> GetCheeseById(int cheeseId)
         {
-            return await _cheeseRepository.GetCheeseById(cheeseId);
+            if (cheeseId <= 0)
+                throw new ArgumentOutOfRangeException();
+
+            var cheese = await _cheeseRepository.GetCheeseById(cheeseId);
+            return cheese;
         }
 
         public async Task<IEnumerable<CheeseDTO>> GetCheeses()
@@ -27,11 +34,17 @@ namespace CheeseApp.Server.Services
 
         public void UpdateCheese(int cheeseId, CheeseDTO cheese)
         {
+            if (cheeseId <= 0)
+                throw new ArgumentOutOfRangeException();
+
             _cheeseRepository.UpdateCheese(cheeseId, cheese);
         }
 
         public void DeleteCheese(int cheeseId)
         {
+            if (cheeseId <= 0)
+                throw new ArgumentOutOfRangeException();
+
             _cheeseRepository.DeleteCheese(cheeseId);
         }
     }
